Major Refactor: ExampleMatchFinder and ExampleFinder
Created an ExampleMatchFinder to define how to search and match a
regular expression to find all the examples strings in a given source.
The ExampleMatchFinder will then say with which ExampleParser the string
should be parsed to extract an Example.

The ExampleFinder will iterate over all the ExampleMatchFinders to find
and parse all the examples.

This combines the resposability of the older ExampleParser and
ExampleMultiParser into one.