Do not print the aborted; instead print the skipped examples
If an example aborts (crashes or the user press ctrl-c), the whole
execution stops, like in a fail-fast scenario. Printing how many
examples aborted is pointless as it can only show 0 or 1.

Instead, change the status to ABORT, making more explicit what had happen.

In replace of the aborts, print the skipped examples.