Init an interactive session if an example fails (experimental)
This experimental feature allows the user to init an interactive session
with the interpreter if an example fails due an output mismatch (time
outs and crashes are not considered).

Regardless of what the user does, the test will still count as a
failure.

This is an experimental feature as it was hard to make the interactive
session to behave like a normal one because some weird behavior between
the interpreter, pexpect and the terminal and it is not clear if the
proposed implementation will work universally.

Another caveat is that if the user doesn't leave the session in a clean
state, it is likely that the rest of the examples fail due mismatches
or timeouts introduced by the user.