The hooks are loaded as modules too.
Like Finders, Parsers and Interpreters, the Hooks are loaded from the
modules with the same mechanism.

Added the possibility of a Hook (and others) to de register itself if the
target (or language) attribute is None.
This check is done by load_modules after instantiating the hook object
(or other) with the configuration as its parameter.
With this, the object can check defaults and command line's parameters
to disable itself or not.