Send one source line and expect one prompt at time
Instead of sending all the source code at the same time, send
one line of code at time.
This is less efficient but it is much closer to what happen when a real
user is behind the keyboard.

Then, after sending a line, wait for the prompt.
The previous code assumed that the code executed was a single expression
and therefor it should only wait for a single PS1 (plus some extra PS1
prompts due the extra newlines added at the end of the expression)

With the introduction of the examples found by Markdown, an example may
be composed by multiple expressions: so multiple PS1 may arise.

Now, we wait for any prompt (PS1, PS2 or whatever) after each line
and we wait for PS1 for the last line.

It is a pending task to redefine the TIMEOUT option.