New flag: CAPTURE; refactor expected regexs engine
The old expected_as_regex (now named expected_as_regexs) will
return a list of regular expressions that would represent the expected
string of the example instead of returning a single regex.

This should enable us to perform an incremental match when comparing the
expected with the got output making the diff a little more cleaner and
easier to spot.

The _add_as_regex also was modified trying to split the literals in
lines, creating a regex per line. This again should improve the diff
algorithms.

A new flag was added: CAPTURE. If true (the default) it will consider
the <...> strings as capture tags but if it is false, it will consider
them as literals.