Calculate the position in expected from where the regex are built.
Include not only the regexs that form what the user is expecting as a
result of a given example but also include the position of each regex
in the expected string from where it was build.

This will be the base for more cleaner and smarter diff algorithms.

Improved also the misaligned message.