Relaxed check_example_overlap
If two examples overlap partially and one is fully included in the
other, do not raise an exception and assume that the larger is the
correct (the smaller or inner is dropped).

If two examples are of the same length and of the same language do not
raise an exception either.

In others cases, raise an exception.