Fix multi issue bug in the expected_as_regexs with WS
When the WS flags was on (whitespace), the expected_as_regexs returned
an invalid list of positions.

First, the WS triggered the deletion of all the whitespaces removing the
vital \n characters making the expected_as_regexs (_add_as_regex) to
work suboptimal.

Second, the _add_as_regex was calculating the positions incorrectly due
a "typo": the line to calculate the charno was below the 'if' structure
instead of being above.

Third, the expected_as_regexs was adding a \n? regex but it should be
\n* to work correctly with and without the WS flag.