Do not allow to generate the pattern \s+\s+ (really slow)
The \s+\s+ is a case of a pathological regular expression. The string
of, 8 spaces could be captured lisk 1-7 spaces, 2-6, 3-5, ... 7-1 spaces
which gives 7 combinations (linear).
If we have \s+\s+\s+ then things starts to grow faster and the regex can
be **really** slow.

The modified _add_as_regex should prevent this. Also, the non-capture
version of the <...> should improve too.