Reimplemented the WS mode and partial matching (exp)
Optimized the regexs built in WS mode to be line oriented (to support
incremental matching), to ignore any whitespace and to make the capture
tag to capture the leading/trailing whitespace if it is not surrounded
by whitespace (it will capture it if not)

A partial matching tries to match as much as possible to generate an
expected string closer to the got, making the diff between both smaller
and more meaningful.

For this we implemented the rcount or real count and the regex's
positions.

A 'real count' is how many literal are in a line. In normal
circumstances, rcount == len(line) (including the new line) but if
normalize_whitespace is True, a sequence of whitespaces count just as
+1.

A regex's position is the position in the original expected string from
where the regex was built.