Improve logs and fixed issue in the end regexs.
Save in the Example the Finder instance that found it. This if loging
purposes later.
Also, build and return an ordered list of the capture tags to keep the
same order in which they were found. Again, this improve the logs.

Fixed an issue when the capture was the last thing in the expected (the
string of literals between the capture and the end is empty). In this
case, the capture's regex must not capture a trailing newline (or a
trailing whitespace in normalize_whitespace==True): those are already
consumed.