For a repeated tag, count it as +1
For the first capture tag <foo>, the tag doesn't give any real count
(because the tag can match anything).
But for the second capture tag <foo>, this gives a rcount of +1 because
it will not match anything but a previous string.
We cannot set a rcount +n (the length of the previous string).