Fixed the u/b Python's markers
Strings like 'b' were replaced by '' because the b' part was confused by
the engine.
The new implementation fixes this with a more precise regex to match the
u/b markers.
In the other hand, however, this requires to patch the _safe_repr
function of pprint. A hack the it works... for now.