Created an OptionParser and use it to parse Example's options
Instead of parsing ourselves the options of an example, we create an
OptionParser to do it based on the argparse.ArgumentParser.

This is a disruptive change but it simplify the code and allows more
advanced things.

At the same time, instead of having a default Options, we have a default
OptionParser and the default options are created from parsing the empty
string using the default OptionParser.

Then, we can extend this parser from a Parser. This will allow the user
to add flags only for a specific language.
With this, we can extend the default options parsing the example's
option string.

This changed also force to change how we choose the diff algorithm. The
keywords UDIFF, NDIFF and CDIFF are gone. Instead we use DIFF and
instead of being a boolean flag, it is an options with an argument with
the same possible values like in the command line.