Options are now lowercase; WS is now norm-ws.
This was the original intent: all the options of byexample should be
lowecase like "+capture" instead of "+CAPTURE".

The uppercase version was to be compatible with Python's doctests but
this tainted the rest of the system and other languages.

Now that we have the possibility to change the parser for the example's
options, we could create a specific python-parser to make a
compatibility layer.