OptionParser pass kwargs to ArgumentParser
In particular, do not add a help for the default option parser.
Otherwise if anyone wants to extend it, it will have a conflict due the
-h flags of both parsers.

For this reason, the default extend_option_parser method does not return
the parent parser but creates an empty one to 'extend' the parent
parser (enabling the help).