Do not add a default to the parser (introduces unwanted names)
If a parser has an option with a default X, after parsing, the option
will be added to the namespace even if it is not present in the parsed
string.

In this case it will have the value X. But Options is an object that it
is designed to handle missing values. Adding X it will cover the values
for the same option below.

Setting argparse.SUPPRESS as a default eliminates this. If an option is
not present nothing is added to the namespace. Then, if we want to get
the value of that option in Options, it will look it through all the
layers until find the first value (or raise an exception if nobody added
it to Options).