Mask/unmask a default value for Options
If a key is not found, the default is to raise a KeyError.
But if the mask_default is called, a default value is returned.

Several calls to mask_default add layers of default values, returning
the last if a key is not found.
The stack can be pop-ed calling unmask_default.