Fixed repr of Options; updated its tests
The repr (__repr__) of Options was calling repr() over a dictionary.
It was correct but it makes the test unpredictable because a repr of a
dict has the keys unordered.

Instead, we use pprint.pformat to make sure that the keys are sorted and
for large Options, we should get a better result.