Made the byexample to work like the doctest's exception handling
doctest detects a traceback and compares its message with the got
string.
byexample, in doctest compatibility mode, it will do almost the same.

In addition, it has a relaxed check: in Python 2.x the exception message
starts with a class  name while in Python 3.x starts with a full name
(module dot class name)

This breaks almost all the exception checks in doctest so we fixed this.