Fixed the empty line issue (which breaks the python indentation)
See doc/languages/python for a complete description of the issue but in
a nutshell, an empty line is interpreted by python as the end of the
previous indented block.

So things like:
 >>> def f():
 ...
 ...    return

will not work out of the box. The old solution was to add enough
trailing spaces to the empty lines to match the expected indentation.

Ugly.

Now, byexample see this and remove any empty line if it is followed by
an indented line, avoiding the close of the indented block.

The only catch is
 >>> s = '''
 ...
 ...    foo'''

where the legal empty line is removed. In these cases, the flag
py-remove-empty-lines can disable the fix.