Allowed empty lines to be non-indented.
Otherwise, the user needs to add trailing spaces to keep everything
aligned.
When using with a prompt it is ok:
  >> a
  ..
  .. b

but with the markdown fenced code, it is ugly
  ```bla
  a

  b
  ```

Note that this also open the question. Are the tasks (remove prompt,
remove indentation) a Parser's responsibility or it should be a Finder's
one.