Moved the prompt remotion to MatchFinder; created process hook in Parser
This a major refactor.

Originally, Parser.source_from_snippet was used to clean up the prompts
but if the example has prompts or not it would depend of the finder whom
found it.
Therefore, the clean up is done now by the MatchFinder.

This includes:
 - check_and_remove_indent: if the indention is important or not it will
   depend of where it was found.
 - check_keep_matching: as it is health check for
   check_and_remove_indent, this was also moved to MatchFinder.
 - get_snippet_and_expected: this is the hook to clean up the snippet
   and or the expected strings (replaces Parser.source_from_snippet and
   Parser.expected_from_match)

In the other side, the Parser changed its interface as follows:
 - the check_and_remove_indent, check_keep_matching, source_from_snippet
   and expected_from_match are gone
 - the Parser.get_example_from_match now is named build_example; it has
   no knowledge about the match (the Example object has no longer a
   match attribute)
 - the subclasses can extend Parser.process_snippet_and_expected to do
   something with the snippets and expected strings (see the Python
   module that uses this to 'mutate' the expected string based on the
   doctest flags).

As stated before, the Python and the Shell modules were updated.
The other modules and some of the documentation are pending.