Explicit delaybeforesend flag; default to None
The delaybeforesend flag is used by pexpect to set a small delay before
sending anything to the underlying program.
This is because most of the programs (interpreters) are designed to
interact with a real human being behind a real keyboard. In such
scenarios, there is an implicit delay before sending any key stroke.

In some applications, the delay is really needed. See the pexpect
documentation for more info.

By default, pexpect set this to 0.05 secs, byexample set this to 0.01.
However this is too slow.

Now, byexample set this to 0 (none) which disables this delay. If
needed, the user can set this to another value globally or per example
using the +delaybeforesend flag.