Reimplemented (and fixed) the universal newlines
First, do not use a regular expression, str.replace should be enough.
Second, do not only map \r\n to \n, but also \r to \n: this is a truly
universal new line.

In addition, remove adding a trailing new line: it is not necessary any
more.

Note: is any other way to transform a string to use universal newlines?
May be a standar function of Python?