Capture all the exception and enhance it (or suppress it)
If verbosity == 0, do not print a traceback. If quiet is True, do not
print anything.
If the exception has a 'where' attribute, print it, otherwise print a
default string.

These three things should improve how the errors are shown to the use,
trying to not print things with low value (traceback) and trying to give
as much as context as possible