Don't use build_exception_msg to enhance the exceptions
build_exception_msg extends or improve the exceptions' messages adding a
'where' and 'who' (or 'owner') strings to give more context (where it
happend like parsing in which file or which lines; who was working like
PythonParser, PythonPromptFinder)

Adding this for each exception, each time is too expensive (duplicate
code) and force to everyone to know the 'where' and the 'who' objects.

Instead, let the exceptions alone (we will enhance them later).