Calculate the capture indexes and flag if any is advanced
The index of each capture regex is the index in the capture
regex list. This can be used to separate regex for capture from
the regex built from the literal strings.

The advanced flag will be true if we are using named capture tags
repeated. That's it, expected likes 'aa<foo>bb<foo>'.

If the expected does not have this special capture tags, the expected
is considered non-advanced.

This discrimination allows to perform further simplifications and
optimizations *only* if the expected is *not* advanced.