Implemented a linear check function (non repeated capture only)
If the expected has no repeated named captures, that it, the underlying
regexs does not contain any back-reference and all the capture tags are
of the form .* (currently this is guaranteed), then we can use a linear
algorithm to check if the expected matches the got or not.

This protect us from pathological regexs with a terrible performance.