Checker refactor: LinearChecker and RegexChecker
Extract the linear and the regex based checkers outside of the Checker
class. This is the first wave of a major refactor.

The RegexChecker will not be used in the practice because the use of
advanced captures is not supported anymore but it is a good opportunity
to refactor the code anyways.