Add the exception's class name and a tip of how to get more info
Add  the exception's class name to the error message. Some errors relay
on the class name to give context.

For example a 'KeyError: x' obviously says that x is a key missing but
if we remove the class name, the message 'x' is not enough.

In addition we added a comment of how to run the example in verbose
mode. An user may no know how to do this and if he wants to debug this
little tip is a good start.