Renamed the Checker/checker by Differ/differ
Now that the check is done by the example's expected, the Checker class
makes no sense as 'checker'; it is just a 'differ'.
Because of this we replace all the Checker/checker instances by
Differ/differ.

Python is a dynamic language so this should not have a direct
backward compatibility issue.
However because this affects the Concern interface, we may declare a new
Major version for byexample (and a minor version for huff)