Fixed regexs when tag was wrapped by whitespace in norm-ws mode
Originally for ' <foo> ' the regexs in norm-ws mode were:
  \s+  (?!\s)(?P<foo>.*?)(?<!\s)  \s+

but if foo maps to empty, the \s+ on the left gets into conflict with
the (?<!\s): \s+ means at least one ws and (?<!\s) means that the
previous char cannot be a ws.

The same contradiction happens with the (?!\s) and the \s+ on the right.

To fix this we replaced the regexs by:
  \s+  (?:(?!\s)(?P<foo>.+?)(?<!\s)\s+)?

The .* was replaced by .+ forcing to match at least one.
If that happen, the \s+ on the right is matched too.
Otherwise the whole group doesn't match (so foo gets empty) and the \s+
on the right isn't executed but still the whitespace is consumed by the
\s+ on the left.

This also avoids a \s+ \s+ pathological condition.

The trailing whitespace of the whole expected and the charnos positions
are special cases.