Set a limit of time for the partial matching
It is a basic implementation where we count how much time we spent doing
the partial matching and aborting if we exceed a threshold.

This will not interrupt a matching in progress (re module has no support
for this) but at least we can detect, eventually, if the process as a
whole is taking to much time and abort it.