Add a new 'rm' option to remove a set of chars
Sometimes you need to remove spurious bytes from the got: rm will do
this; it will remove a set of chars from the expected and the got
strings.

This can be used to fake an empty line that otherwise would mark the end
of the expected string at the half.

This doesn't work:
    >>> print("hello\nworld")
    hello

    world

But this will do the trick:
    >>> print("hello\nworld")   # byexample: +rm=~
    hello
    ~
    world

because the ~ will be removed before doing the check.