Enable/disable the printing of the expr's value
By default, autodetect if the example is expecting an expr's value (aka,
=> prompt) and enable/disable the printing of the expr's value based on
it.

This free the user from adding ; at the end of each line or using
"+pass" hacks.

The user can force a 'always enabled' or a 'never enabled' setting from
the command line.

This change in theory *should* be backward compatible.