Use .. as secondary prompt for ruby
This is a non backward compatibility change.

Before, one or more *consecutive* primary prompts >> were considered a
single example.

This is too fragile and dangerous.

Consider the following (previous this commit)

>> puts "a"     # byexample: +pass
>> puts "b"

Which output is none: the +pass is affecting the 2-lines example.

Now, if you add a single newline (probably for aesthetics purposes)

>> puts "a"     # byexample: +pass

>> puts "b"

The example is split in 2 and the second is has no the +pass option
enabled.

For this reason now each >> is a single example; for multiline examples
we need to use ..

We go from this

>> def f
>>   1
>> end

to this

>> def f
..   1
.. end