Two hooks (end_example, finally_example); allow got modification from concern
Created a 'meta' attribute in the example, a mutable dictionary. Then,
the executor will save the string got from the example execution there
allowing the concerns to modify it.

For this, two additional hooks were added: end_example (at the end of
the example normal execution) and finally_example (at the end of the
example execution regardless of how it ended).