Allow a Ruby => following by the end of the string.
If => is not followed by anything else, it was interpreted as literal
and it was not activating the auto print expression in Ruby examples.

But this syntax may arise:
 - if the user wants to activate the auto print but it has nothing to
check (he is playing)
 - if the output is put in the line below which is perfectly possible if
+norm-ws is enabled (see the example in this commit)

So now, if => is followed by the end of the line, it will activate the
auto print.

This is a bug fix that may introduce incompatibility changes in the
examples.

Things like:

>> puts "=>\nfoo"
=>
foo

Where possible but now they will trigger the auto print so you will need
to update the expected to something like:

>> puts "=>\nfoo"
=>
foo
=> nil