Delay the examples' build until the last minute
Instead of building all the examples before its execution, build each
one just a second before executing it.

This is a non-backward compatible change that changes the semantics of
Concern.start_run and the Runner.initialize.

In addition, renamed several methods now that they do not work with
examples but they work with example matches.