Integrated the ExampleMatch concept into Example
The ExampleMatch is not used any more and in its place we will use
Example.

This last one will have two states: before and after the parsing phase.
During the first we will play the role of the old ExampleMatch.

Created a auxiliar func _build_fake_example to help the testing.