Removed Example's meta attr; write directly to the example
The attribute meta was a hack to write in a non-writeable example
object.

But now it is possible to write on it directly so the meta is not needed
anymore.

For example, the executor will save in example.got the string got from
the execution instead of saving it in example.meta['got'].