Removed +shell option from Shell runner
This is a non-backward compatibility change that remove the +shell
option. This one allowed the user to change the shell from sh to bash
(or even other).

But after the introduction of --shebang, the use case of +shell was
almost none and worst, the current implementation was compatible with
the --shebang so removing +shell is the correct decision.