Fix: Do not replace - by _ in the capture tag.
The replacement of - by _ was an internal detail that didn't affected
the semantics of the named capture tags until the implementation of the
clipboard module (+paste option).

This required that a capture tag <foo-bar> were used as <foo_bar> in
paste mode.

This commit fixes this so <foo-bar> can be used in paste mode as
expected.