Use _ instead of - for the regexs' names
Map the tag names (that could have -) to a regex name replacing the - by
the _ just before building the regex.
Perform the inverse when retrieving the captured texts from a regex
match.

This enables us to use tags like <foo-bar>, which sound less robotic than
<foo_bar>, integrated with the Python regex engine that request names
like (?P<foo_bar>...).

Any tag name that is visible to the user should be <foo-bar> and not
<foo_bar> after this commit.