Document how to and the limits of running in background.
byexample waits for the completion of the example so running
in background will yield immediately but it may print things to
the terminal asynchronously which would interfere with the output of the
subsequent examples.

This doc explains how to work with processes in background and the
limits and trade-offs of each strategy.