The unnamed tag's regex is greedy.
This potentially non-backward compatible change sets that the unnamed
tag's regex (for <...>) is greedy (.*) in contrast with the named tag's
one (.*?)

This is an heuristic: the <foo> is used to capture interesting but not
deterministic outputs, typically small strings while the <...> is used
for uninteresting outputs that may be small strings but they could be
also long sentences or even they could span multiple lines.

The greedy mode fits more appropriately.