If an example fails in fail-fast mode; skip the rest unless said -skip
If fail-fast is on (because the user set --ff from the command line or
because he set +fail-fast from the example's option) and if the example
fails due an unexpected output, the rest of the examples are skipped.

This differ a little from the previous behaviour because before, the
"skipped" examples were literally avoided. Now, the skipped examples are
still parsed but not executed (the skip_example hook is called)

A given example can force its execution however, using -skip flag.

If the example fails for other reason (crash or timeout) the old
behaviour is used (even in presence of -skip).
This is because we cannot relay in the state of the interpreter
anymore so we need to fail fast *hard*.

This commit enables a basic setup and tear down functionality.