Set delayafterread to None for performance reasons
This change prevent Python to release its GIL when looping in the
pexpect's read operations.

But due that byexample is not multithreading (and it is very unlikely),
releasing the lock is not an issue.

Setting delayafterread to None speed ups the execution (in a crude
experiment, a run that took ~8 secs now takes ~7 secs).