Do not count the last newline if it doesn't exist
This fix a small bug. It was assumed that the example
always ended in a newline but this char is optional in the finders.

Because that newline is not wanted the code did a '-1' to the counter
but if the example doesn't end in a newline, that would be removing
a valid newline from the counter.

Now the optional newline is removed directly. (It is easier to do
this than fixing the finders' regexs).