Fix: make optional the capture at the end if it's preceded by a newline
In normal mode (normalize_whitespace = False), the last tag regex that
is at the end of the expected will have a (?<!\\n) regexs that means
"not preceded by a newline".
But if the capture matches the empty string and it is preceded by a new
line, the (?<!\\n) will always fail.

To prevent this make this regex optional.

Fixes #51.