Remove spurious endings like ``` at the end of an example.
By default if an example ends with one of ```, ''', ~~~ or -->, the last
line will be removed as it is assumed that those tokens are not part of
the example but just they are the product of some mark-up that wraps the
example.

The method ExampleFinder.remove_spurious_endings performs this remotion.

ExampleFinder.spurious_endings returns a set of the tokens that are
considered spurious; a subclass can override it removing some of them if
they affect how the particular finder finds an example.

See FencedMatchFinder subclass for an example of this.

This commit closes #47, it is backward compatible but it is quite
experimental.

Currently the initial set of tokens is hardcoded and cannot be changed.