Change default terminal's geometry (24x80) from cmdline
The default geometry or dimension used was the ptyprocess's default: 24
lines and 80 columns.

Now, this can be changed from command line (from an example is not
supported).

In addition to this, the LINES and the COLUMNS enviroment variables are
computed and forwarded to the spawned processes.

This is backward compatible change as the geometry by default is the
same and the enviroment was forwarded by default too.

Note: this may not have any effect on how the the runners print.
For example Python's pretty printer prints with a fixed width of 80
chars/columns (to be done).

Closes #34.