Support report progress in parallel.
Avoid any race condition with the output of multiple reporters writing
to it concurrently; however other prints may interleave with the
reporters' ones as they are not synchronized.

Set tqdm to display a progress bar for each independent reporter.

If only one report is executed, avoid any penalty of synchronization.