Executor return 3-uple: fail? cancelled? user aborted?
This tuple allows to differentiate if the example suite failed or it was
cancelled due a crash or a timeout or if it just was the user who
aborted the execution.

A fail implies FAIL, a cancel or an user abort
implies an ABORT.

If --ff is not enabled, the first and the second cases don't abort the
whole execution of other files; it --ff is on, it cancels the whole
execution.