Use terminal emulator to handle cling's output.
cling doesn't disable its TTY output which makes its output
a really hard to parse one, specially with the inputs echoed back.

There was a fix before that filtered out all the terminal escape
opcodes but it was too fragile.

A cling's syntax error was the case because the error contains text in
color which obviously means that the text captured by byexample
contained escape codes in over all the place.

pyte is a ANSI Terminal emulator that is put in between the pexpect
captured output and the byexample's engine.

This filter out uninterested escape codes like the color and simplify
the filtering of the echoed lines.

Added an example of how to capture and test a cling syntax error.