Split cache's activate and sync steps into 2 ctx managers
Split the cache context manager into two: one for activate the cache
(patch on enter and unpatch on exit) and the other to sync the cache to
disk (clear the cache's stats on enter and sync if needed to disk on
exit).

This simplified the code and enabled to open and load the cache just
once, avoiding if possible any hit to the disk.