Toggle IRB echo in the same line that the cmd to execute
In Ruby 2.3.3, "A;\nB" was equivalent to "A;B" but in Ruby 2.5.3 (and
probably before), the IRB console will print the value of the expression
in a line regardless if it ends in a semicolon or not.

byexample adds a ruby code before executing the example to toggle the
IRB echo mode (see +ruby-expr-print).

With the change in how IRB prints the expressions, we need to make sure
that the toggle code is in the same line than the cmd.

This fix is not bullet-proof: if the cmd starts with a newline, it will
trigger the bug again.

A more robust solution could be execute the toggle code in its own
_exec_and_wait loop, discarding its output but each _exec_and_wait is
expensive so the decision to keep sending the toggle code and the cmd in
one _exec_and_wait is a tradeoff between robustness and speed.

This closes #62.