Refactor PexepctMixin/ExampleRunner passing options explicitly.
Pass to _exec_and_wait, _expect_prompt, _change_terminal_geometry_ctx,
_change_terminal_geometry and _get_output the options
instead of passing some of them by arguments. This breaks the
compatibility but it will enable further extensions (via options) in a
painless way.

For some methods, some explicit arguments are still allowed and take
priority over the values in options due that those are typically changed
frequenctly like 'timeout' of _expect_prompt or 'rows' of
_change_terminal_geometry.

The ExampleRunner.run also changed in a less disruptive way: its 'flags'
parameter was renamed to 'options' to make it more consistent with the
rest of byexample.