Force +term-emu and disable +geometry for CPP examples
The cling interpreter's output needs to be post-processed by a ANSI
Terminal emulator so +term-emu is always on for CPP examples.

The +geometry is disabled because the output gets mixed with a lot of
echos from the terminal resize if the terminal is too small and this
breaks _get_output.