Support 3 terminal modes: as-is, dumb (default), ansi
Previous this commit the default terminal emulation mode worked like the
'as-is' mode: the new lines of the interpreter's output are standardized
and nothing else.

This mode is still available but the default changed to a new mode: dumb
terminal mode.

This, in addition to the universal new line treatment it strips any
trailing whitespace and replace any tab by spaces emulating a real
terminal (with a tab-size of 8 spaces). No other emulation is done.

This new default is not backward-compatible of course, but it will aim
to simplify the tests and reduce the use of +norm-ws.

Finally the ansi mode emulates a full ANSI terminal.

The modes can be changed using +term (+term-emu was removed).

This closes #41.