Build a regex for a tag.
This concentrate the code that generates the regex in one single place
and simplify the possibilities.

In particular, if the regex has not to consume whitespace/newlines
before or after, the whole regex is made optional always.

For 'a<foo>', the whole regex before was r'\Aa(?P<foo>.*?)(?<!\\n)\\n*\Z'
but now it is r'\Aa(?:(?P<foo>.+?)(?<!\\n))?\\n*\Z'.