Fix 2 bugs: never begin with a regex; allow nl except on last
Never begin a regex chunk with a non-literal regex like '\s+'. This
breaks the assumption in _LinearExpected that each chunk begins with a
literal. This of course does not apply to the tags.

If this assumption is broken the following in norm-ws 'a <foo> b' is
translated to: ['a\s+', (tag regex here), '\s+b']. Then the
_LinearExpected will see: ['a\s+', '\s+b']; when the first regex
matches, it will consume all the whitespace and the second will always
fail (because _LinearExpected has no backtracking)

The solution is ['a\s+', (tag regex here)\s+?, 'b'] but this has the
problem that the string 'a b' is considered correct when the expected
request for at least 2 spaces. This limitation was precent in 7.4.4 too.

The second issue involves that the posfix (?<!\\n) must be added only if
the tag is followed by a \\n*.

Otherwise '\n<foo>\n' fails to match '\n\n\n' because the (?<!\\n)
prevents that <foo> has a newline.