Unamed tags non greedy by default except at endline
As the named tags, the unamed tags are non-greedy regexs by default.
This is possibly a non-backward compatible change (the unamed tags were
greedy by default previous this commit).

However there is an exception: if the unamed tag is a the end of a line
(it is followed by a new line), the regex will be greedy.
This exception is based on a heuristic: when a ellipsis is at the begin
or in the middle of a line, typically is used to remove some small
unwanted string in the same line but when it is in the end or in its own
line, the ellipsis is used to remove large multi-line strings.