Select Zone Finder by file extension, use the whole file as default
Instead of processing the whole file, process only the zones or areas
defined by the Zone Finder. Select which zone finder to use based on the
file extension (.txt by default)

If not suitable zone finder is found, rollback to the previous strategy
and process the whole file. This may change in a future.