Remove spurious and specific concepts; impl zone finder for Python files
Removed the concept of spurious endings and specific/generic finders.
These two were implemented to remove unwanted strings from a too
free-form finder and to avoid overlapping between examples due the
Markdown fenced-code finder and other more 'specific' finder.

These two problems were removed: the free-form can be controlled with a
Zone Finder (implemented one for Python in this commit) and the
overlapping is not an issue anymore as the Markdown fenced-code is gone.