The # >> Ruby prompt is not supported.
The Ruby prompt >> was extended to support the # >> notation because it
was thought that a Ruby comment (those which starts with #) would be the
perfect place to put examples.

But now that we support zone finders, the Ruby comments are found and
processed before even trying to find an example so the Ruby examples
using the >> prompt will be found in any zone including the Ruby
comments without needing a special prompt.