Set the default of every option in its optparser.
get_options() (init) will not hardcode the default values for
its options anymore. Instead, it sets the default in the OptionParser
itself.

The same is done for all the OptionParsers of all the modules.

This default values are *not* used during the parsing but they are
collected *before* by extend_option_parser_with_concerns,
extend_options_with_language_specific and get_options functions (init
module).

The collected values are saved in the Options stack *before* the
parsing.

This allows to set the defaults once and on the bottom of the stack. If
OptionParser would set the defaults during the parsing, the defaults
would be too high in the stack and they *could override* the values set
before.

The whole purpose is to have a more simpler interface: now each
ExampleParser can add options and set their default without worrying
about the *override* problem and it can be sure that a default value
will exist in the options variable/parameter.