Test several ruby version under Travis CI
Keep running all the tests in different versions of Python but also run
some tests (ruby related) in different versions of Ruby with a version
of Python fixed.

Currently we tests Ruby 2.0, 2.1, 2.2, 2.3, 2.4, 2.5 and 2.6. Modern
versions of Ruby (like 2.7) may be supported as well but Travis CI
currently does not have the infrastructure to run under those
environments.