Recover the runner after a timeout and do not abort.
If possible, try to recover the control of the runner after a timeout.
On success, mark the example as failed and not as timedout (but still
call the appropriated concern hook).

This enable us to call the rest of the examples, specially the ones that
perform some clean up.

However this feature depends on the implementation of each runner.

If the recovery is not possible, rollback to the previous behaviour and
abort the whole execution.

If needed, this can be forced with the flag -x-not-recover-timeout.