Verify output encoding and exit earlier if it fails
Python 2.x uses the terminal encoding for its stdout. But if the stdout
is redirected to a pipe or file the encoding is set to None, which
defaults to 'ascii'.

This breaks the output of unicode examples.

It is hard to fix this from byexample itself so the solution is to
verify the stdout encoding and if it is None print a message and exit.

The only fix for this is using the PYTHONIOENCODING env variable.

The problem only affects Python 2.x when the output is redirected to a
pipe or file.

Closes #80