Set bash as the default shell instead of sh
The 'sh' shell is more than ambiguous: in Red Hat it is a link to bash,
while in Debian 9 it is a link to dash (and previously, around Debian 6
was a link to bash).

To avoid this ambiguity, the new default is bash, a shell that it is
available and installed by default in most distros.

Depending of the environment this can or not be backward incompatible:
in Red Hat based distro  and any one who has 'sh' as a link to bash this
change is transparent.

For modern versions of Debian, Ubuntu and others this could break some
examples:
 - the 'echo' program *may* have different default options. You *may*
need to add them explicitly if needed ('echo' in dash has the '-e' flag
on by default while in bash don't: you need to replace 'echo' by 'echo
-e' explicitly)
 - some symbols *may* need to be escaped (like '!' which in bash has
a special meaning)
 - the job control/monitoring output *may* change (see
docs/languages/shell.md for a full explanation of the differences).

Nevertheless this commit introduces also a shortcut for the '-x-shebang'
functionality: the '+shell' option.

With this, the user can switch back the underlying shell to 'sh'
('+shell=sh').